+++ auto_route/lib/src/router/controller/routing_controller.dart
@@ -17,8 +17,7 @@ import '../../utils.dart';
 part '../../route/route_data.dart';
 
 typedef RouteDataPredicate = bool Function(RouteData route);
-typedef OnNestedNavigateCallBack = void Function(
-    List<RouteMatch> routes, bool initial);
+typedef OnNestedNavigateCallBack = void Function(List<RouteMatch> routes, bool initial);
 typedef OnTabNavigateCallBack = void Function(RouteMatch route, bool initial);
 
 abstract class RoutingController with ChangeNotifier {
@@ -34,8 +33,7 @@ abstract class RoutingController with ChangeNotifier {
     _childControllers.remove(childController.routeData.key);
   }
 
-  List<RouteData> get stackData =>
-      List.unmodifiable(_pages.map((e) => e.routeData));
+  List<RouteData> get stackData => List.unmodifiable(_pages.map((e) => e.routeData));
 
   bool isRouteActive(String routeName) {
     return root._isRouteActive(routeName);
@@ -52,8 +50,7 @@ abstract class RoutingController with ChangeNotifier {
     List<RouteMatch>? pendingRoutes;
     final routeKey = routeToPush.key;
     if (_pendingInitialRoutes.containsKey(routeKey)) {
-      pendingRoutes =
-          List<RouteMatch>.unmodifiable(_pendingInitialRoutes[routeKey]!);
+      pendingRoutes = List<RouteMatch>.unmodifiable(_pendingInitialRoutes[routeKey]!);
       _pendingInitialRoutes.remove(routeKey);
     } else if (routeToPush.hasChildren) {
       pendingRoutes = routeToPush.children!;
@@ -79,8 +76,7 @@ abstract class RoutingController with ChangeNotifier {
         onFailure(RouteNotFoundFailure(route));
         return null;
       } else {
-        throw FlutterError(
-            "[${toString()}] Router can not navigate to ${route.fullPath}");
+        throw FlutterError("[${toString()}] Router can not navigate to ${route.fullPath}");
       }
     }
   }
@@ -108,21 +104,16 @@ abstract class RoutingController with ChangeNotifier {
   }
 
   bool _isPathActive(String pattern) {
-    return RegExp(pattern)
-        .hasMatch(p.joinAll(currentSegments.map((e) => e.stringMatch)));
+    return RegExp(pattern).hasMatch(p.joinAll(currentSegments.map((e) => e.stringMatch)));
   }
 
   bool _canHandleNavigation(PageRouteInfo route) {
     return routeCollection.containsKey(route.routeName);
   }
 
-  _RouterScopeResult<T>?
-      _findPathScopeOrReportFailure<T extends RoutingController>(String path,
-          {bool includePrefixMatches = false, OnNavigationFailure? onFailure}) {
-    final routers = [
-      if (this is T) this as T,
-      ..._getAncestors().whereType<T>()
-    ];
+  _RouterScopeResult<T>? _findPathScopeOrReportFailure<T extends RoutingController>(String path,
+      {bool includePrefixMatches = false, OnNavigationFailure? onFailure}) {
+    final routers = [if (this is T) this as T, ..._getAncestors().whereType<T>()];
     for (var router in routers) {
       final matches = router.matcher.match(
         path,
@@ -144,23 +135,19 @@ abstract class RoutingController with ChangeNotifier {
     return null;
   }
 
-  RoutingController _findScope<T extends RoutingController>(
-      PageRouteInfo route) {
+  RoutingController _findScope<T extends RoutingController>(PageRouteInfo route) {
     if (_parent == null || _canHandleNavigation(route)) {
       return this;
     }
     final routers = [this, ..._getAncestors()];
-    return routers.firstWhere((r) => r._canHandleNavigation(route),
-        orElse: () => this);
+    return routers.firstWhere((r) => r._canHandleNavigation(route), orElse: () => this);
   }
 
-  Future<dynamic> navigate(PageRouteInfo route,
-      {OnNavigationFailure? onFailure}) async {
+  Future<dynamic> navigate(PageRouteInfo route, {OnNavigationFailure? onFailure}) async {
     return _findScope(route)._navigate(route, onFailure: onFailure);
   }
 
-  Future<dynamic> _navigate(PageRouteInfo route,
-      {OnNavigationFailure? onFailure}) async {
+  Future<dynamic> _navigate(PageRouteInfo route, {OnNavigationFailure? onFailure}) async {
     final match = _matchOrReportFailure(route, onFailure);
     if (match != null) {
       return _navigateAll([match], onFailure: onFailure);
@@ -188,8 +175,7 @@ abstract class RoutingController with ChangeNotifier {
   }
 
   List<RoutingController> _getAncestors() {
-    void collectRouters(
-        RoutingController currentParent, List<RoutingController> all) {
+    void collectRouters(RoutingController currentParent, List<RoutingController> all) {
       all.add(currentParent);
       if (currentParent._parent != null) {
         collectRouters(currentParent._parent!, all);
@@ -270,8 +256,7 @@ abstract class RoutingController with ChangeNotifier {
   @override
   String toString() => '${routeData.name} Router';
 
-  Future<void> _navigateAll(List<RouteMatch> routes,
-      {OnNavigationFailure? onFailure});
+  Future<void> _navigateAll(List<RouteMatch> routes, {OnNavigationFailure? onFailure});
 }
 
 class TabsRouter extends RoutingController {
@@ -391,8 +376,7 @@ class TabsRouter extends RoutingController {
   }
 
   @override
-  Future<void> _navigateAll(List<RouteMatch> routes,
-      {OnNavigationFailure? onFailure}) async {
+  Future<void> _navigateAll(List<RouteMatch> routes, {OnNavigationFailure? onFailure}) async {
     if (routes.isNotEmpty) {
       final mayUpdateRoute = routes.last;
 
@@ -413,13 +397,11 @@ class TabsRouter extends RoutingController {
           if (mayUpdateController.managedByWidget) {
             if (mayUpdateController is StackRouter) {
               mayUpdateController.onNavigate?.call(newRoutes, false);
-            } else if (mayUpdateController is TabsRouter &&
-                newRoutes.isNotEmpty) {
+            } else if (mayUpdateController is TabsRouter && newRoutes.isNotEmpty) {
               mayUpdateController.onNavigate?.call(newRoutes.last, false);
             }
           }
-          return mayUpdateController._navigateAll(newRoutes,
-              onFailure: onFailure);
+          return mayUpdateController._navigateAll(newRoutes, onFailure: onFailure);
         } else {
           final data = _createRouteData(mayUpdateRoute, routeData);
           _pages[pageToUpdateIndex] = pageBuilder(data);
@@ -450,8 +432,7 @@ class TabsRouter extends RoutingController {
   @override
   bool get canPopSelfOrChildren {
     if (_childControllers.containsKey(_pages[_activeIndex].routeData.key)) {
-      return _childControllers[_pages[_activeIndex].routeData.key]!
-          .canPopSelfOrChildren;
+      return _childControllers[_pages[_activeIndex].routeData.key]!.canPopSelfOrChildren;
     }
     return false;
   }
@@ -488,8 +469,7 @@ class TabsRouter extends RoutingController {
       fragment: fragment,
     ));
     if (includeAncestors && _parent != null) {
-      _parent!
-          ._updateSharedPathData(queryParams: queryParams, fragment: fragment);
+      _parent!._updateSharedPathData(queryParams: queryParams, fragment: fragment);
     }
   }
 }
@@ -545,8 +525,7 @@ abstract class StackRouter extends RoutingController {
   bool get canPopSelfOrChildren {
     if (_pages.length > 1) {
       return true;
-    } else if (_pages.isNotEmpty &&
-        _childControllers.containsKey(_pages.last.routeData.key)) {
+    } else if (_pages.isNotEmpty && _childControllers.containsKey(_pages.last.routeData.key)) {
       return _childControllers[_pages.last.routeData.key]!.canPopSelfOrChildren;
     }
     return false;
@@ -612,6 +591,14 @@ abstract class StackRouter extends RoutingController {
     }
   }
 
+  @optionalTypeArgs
+  void popForced<T extends Object?>([T? result]) {
+    final NavigatorState? navigator = _navigatorKey.currentState;
+    if (navigator != null) {
+      navigator.pop(result);
+    }
+  }
+
   bool removeLast() => _removeLast();
 
   void removeRoute(RouteData route, {bool notify = true}) {
@@ -644,8 +631,7 @@ abstract class StackRouter extends RoutingController {
   List<AutoRoutePage> get stack => List.unmodifiable(_pages);
 
   @optionalTypeArgs
-  Future<T?> push<T extends Object?>(PageRouteInfo route,
-      {OnNavigationFailure? onFailure}) async {
+  Future<T?> push<T extends Object?>(PageRouteInfo route, {OnNavigationFailure? onFailure}) async {
     return _findStackScope(route)._push<T>(route, onFailure: onFailure);
   }
 
@@ -660,8 +646,7 @@ abstract class StackRouter extends RoutingController {
     );
   }
 
-  Future<dynamic> _popUntilOrPushAll(List<RouteMatch> routes,
-      {OnNavigationFailure? onFailure}) async {
+  Future<dynamic> _popUntilOrPushAll(List<RouteMatch> routes, {OnNavigationFailure? onFailure}) async {
     final anchor = routes.first;
     final anchorPage = _pages.lastOrNull(
       (p) => p.routeKey == anchor.key,
@@ -933,8 +918,7 @@ abstract class StackRouter extends RoutingController {
         if (mayUpdateController.managedByWidget) {
           if (mayUpdateController is StackRouter) {
             mayUpdateController.onNavigate?.call(newChildren, false);
-          } else if (mayUpdateController is TabsRouter &&
-              newChildren.isNotEmpty) {
+          } else if (mayUpdateController is TabsRouter && newChildren.isNotEmpty) {
             mayUpdateController.onNavigate?.call(newChildren.last, false);
           }
         }