+++ auto_route/CHANGELOG.md
@@ -1,4 +1,8 @@
 # ChangeLog
+## [2.2.1] 
+- Fix crash when passing non-string values as query params #606
+- Update analyzer version to 2.0.0
+- Update build_runner,build and source_gen versions
 ## [2.2.0] Breaking changes!
 - Remove helper methods pushToChild and replaceInChild from TabsRouter [Bad idea]. [Breaking]
 - Add navigate and navigateNamed methods to TabsRouter